cli: move portForwarder to a more generic location
This test just moves portForwarder to a more generic/utils
location.

Testing will require immediate use of port-forwarding; moving
it to a more generic and less bloated will reduce unnecessary
imports and coupling.
